feat: add core types for `venula`, and begin work on the typeclass
The types here are meant to provide the user with a graph
accurately representing the data the user is holding in state.
The typeclass gives this all to the user if they can only
define an `Iso'` between their state representation, and a pair
of `Vertex`s parameterized by their own custom sum type.