refactor: simplify venula
Refactor the library to use a user provided state representation.
Now they just tell us how to get the representation from and to
the state via `IO`, and we do the rest. For tests they can provide
either a generator of a pair, or a pair of state representations
that represent the state the tested function should be moving
the program between.